Legacy storage file deprecation
