+++ src/androidTest/java/ImpressionsFileStorageTest.java
@@ -5,32 +5,19 @@ import org.junit.Before;
 import org.junit.Test;
 
 import java.io.File;
-import java.net.URI;
 import java.net.URISyntaxException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import fake.HttpClientStub;
-import fake.SplitCacheStub;
-import io.split.android.client.TrackClient;
-import io.split.android.client.TrackClientImpl;
-import io.split.android.client.dtos.Event;
 import io.split.android.client.impressions.IImpressionsStorage;
 import io.split.android.client.impressions.Impression;
-import io.split.android.client.impressions.ImpressionsFileStorage;
-import io.split.android.client.impressions.ImpressionsManager;
+import io.split.android.client.storage.legacy.ImpressionsFileStorage;
 import io.split.android.client.impressions.ImpressionsManagerConfig;
 import io.split.android.client.impressions.ImpressionsManagerImpl;
-import io.split.android.client.impressions.ImpressionsStorageManager;
-import io.split.android.client.impressions.ImpressionsStorageManagerConfig;
-import io.split.android.client.track.ITrackStorage;
-import io.split.android.client.track.TrackClientConfig;
-import io.split.android.client.track.TrackStorageManager;
-import io.split.android.client.track.TracksFileStorage;
+import io.split.android.client.storage.legacy.ImpressionsStorageManager;
+import io.split.android.client.storage.legacy.ImpressionsStorageManagerConfig;
 
 public class ImpressionsFileStorageTest {
