+++ src/androidTest/java/integration/PauseBGProcessesTest.java
@@ -37,11 +37,11 @@ import io.split.android.client.dtos.Event;
 import io.split.android.client.impressions.IImpressionsStorage;
 import io.split.android.client.impressions.Impression;
 import io.split.android.client.impressions.ImpressionListener;
-import io.split.android.client.impressions.ImpressionsFileStorage;
+import io.split.android.client.storage.legacy.ImpressionsFileStorage;
 import io.split.android.client.impressions.ImpressionsManagerConfig;
 import io.split.android.client.impressions.ImpressionsManagerImpl;
-import io.split.android.client.impressions.ImpressionsStorageManager;
-import io.split.android.client.impressions.ImpressionsStorageManagerConfig;
+import io.split.android.client.storage.legacy.ImpressionsStorageManager;
+import io.split.android.client.storage.legacy.ImpressionsStorageManagerConfig;
 import io.split.android.client.lifecycle.LifecycleManager;
 import io.split.android.client.network.HttpClient;
 import io.split.android.client.network.HttpClientImpl;
@@ -50,8 +50,7 @@ import io.split.android.client.storage.legacy.FileStorage;
 import io.split.android.client.storage.legacy.IStorage;
 import io.split.android.client.track.ITrackStorage;
 import io.split.android.client.track.TrackClientConfig;
-import io.split.android.client.track.TrackStorageManager;
-import io.split.android.client.track.TracksFileStorage;
+import io.split.android.client.storage.legacy.TrackStorageManager;
 import io.split.android.client.utils.Utils;
 import io.split.android.engine.experiments.RefreshableSplitFetcherProvider;
 import io.split.android.engine.experiments.RefreshableSplitFetcherProviderImpl;
@@ -281,7 +280,7 @@ public class PauseBGProcessesTest {
         trackConfig.setWaitBeforeShutdown(1);
         trackConfig.setMaxSentAttempts(3);
         trackConfig.setMaxQueueSizeInBytes(5000);
-        ITrackStorage eventsStorage = new TracksFileStorage(mContext.getCacheDir(), dataFolderName);
+        ITrackStorage eventsStorage = new FileStorage.TracksFileStorage(mContext.getCacheDir(), dataFolderName);
         TrackStorageManager trackStorageManager = new TrackStorageManager(eventsStorage);
         TrackClient trackClient = TrackClientImpl.create(trackConfig, httpClient, eventsRootTarget, trackStorageManager, splitCache);
