+++ src/main/java/io/split/android/client/SplitFactoryImpl.java
@@ -24,10 +24,10 @@ import io.split.android.client.factory.FactoryMonitorImpl;
 import io.split.android.client.impressions.IImpressionsStorage;
 import io.split.android.client.impressions.ImpressionListener;
 import io.split.android.client.impressions.ImpressionsManagerConfig;
-import io.split.android.client.impressions.ImpressionsFileStorage;
+import io.split.android.client.storage.legacy.ImpressionsFileStorage;
 import io.split.android.client.impressions.ImpressionsManagerImpl;
-import io.split.android.client.impressions.ImpressionsStorageManager;
-import io.split.android.client.impressions.ImpressionsStorageManagerConfig;
+import io.split.android.client.storage.legacy.ImpressionsStorageManager;
+import io.split.android.client.storage.legacy.ImpressionsStorageManagerConfig;
 import io.split.android.client.lifecycle.LifecycleManager;
 import io.split.android.client.metrics.CachedMetrics;
 import io.split.android.client.metrics.FireAndForgetMetrics;
@@ -39,8 +39,7 @@ import io.split.android.client.storage.legacy.FileStorage;
 import io.split.android.client.storage.legacy.IStorage;
 import io.split.android.client.track.ITrackStorage;
 import io.split.android.client.track.TrackClientConfig;
-import io.split.android.client.track.TrackStorageManager;
-import io.split.android.client.track.TracksFileStorage;
+import io.split.android.client.storage.legacy.TrackStorageManager;
 import io.split.android.client.utils.Logger;
 import io.split.android.client.utils.Utils;
 import io.split.android.client.validators.ApiKeyValidator;
@@ -177,7 +176,7 @@ public class SplitFactoryImpl implements SplitFactory {
         trackConfig.setWaitBeforeShutdown(config.waitBeforeShutdown());
         trackConfig.setMaxSentAttempts(config.eventsMaxSentAttempts());
         trackConfig.setMaxQueueSizeInBytes(config.maxQueueSizeInBytes());
-        ITrackStorage eventsStorage = new TracksFileStorage(context.getCacheDir(), dataFolderName);
+        ITrackStorage eventsStorage = new FileStorage.TracksFileStorage(context.getCacheDir(), dataFolderName);
         TrackStorageManager trackStorageManager = new TrackStorageManager(eventsStorage);
         _trackClient = TrackClientImpl.create(trackConfig, httpClient, eventsRootTarget, trackStorageManager, splitCache);
