+++ src/main/java/io/split/android/client/impressions/HttpImpressionsSender.java
@@ -11,6 +11,7 @@ import io.split.android.client.network.HttpException;
 import io.split.android.client.network.HttpMethod;
 import io.split.android.client.network.HttpResponse;
 import io.split.android.client.network.URIBuilder;
+import io.split.android.client.storage.legacy.ImpressionsStorageManager;
 import io.split.android.client.utils.Json;
 import io.split.android.client.utils.Logger;
 import io.split.android.client.utils.Utils;
