+++ src/test/java/io/split/android/engine/impressions/ImpressionsStorageTest.java
@@ -21,9 +21,9 @@ import io.split.android.client.dtos.ChunkHeader;
 import io.split.android.client.dtos.KeyImpression;
 import io.split.android.client.dtos.TestImpressions;
 import io.split.android.client.impressions.IImpressionsStorage;
-import io.split.android.client.impressions.ImpressionsFileStorage;
-import io.split.android.client.impressions.ImpressionsStorageManager;
-import io.split.android.client.impressions.ImpressionsStorageManagerConfig;
+import io.split.android.client.storage.legacy.ImpressionsFileStorage;
+import io.split.android.client.storage.legacy.ImpressionsStorageManager;
+import io.split.android.client.storage.legacy.ImpressionsStorageManagerConfig;
 import io.split.android.client.impressions.StoredImpressions;
 import io.split.android.client.utils.Json;
 import io.split.android.fake.ImpressionsFileStorageStub;
